Moved getWorkbookSession into driveitem
